Move FeedMapper to its own file
